requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
fake-useragent==1.4.0
urllib3==2.2.0